+++ lisp/core/cat.el
@@ -27,7 +27,6 @@
   ;; Preparation
   (eask-with-archives "melpa"
     (eask-package-install 'e2ansi))
-  (package-activate-all)
   (eask-msg "")
 
   ;; Start the task